fog/requirements.txt
paho-mqtt==1.5.1
requests==2.26.0
hvac==0.11.2